+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/annotation/Category.java
@@ -1,7 +1,6 @@
 package guru.qa.niffler.jupiter.annotation;
 
 import guru.qa.niffler.jupiter.extension.CategoryExtension;
-import guru.qa.niffler.jupiter.extension.SpendingResolverExtension;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.lang.annotation.ElementType;
@@ -11,7 +10,7 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
-@ExtendWith({CategoryExtension.class, SpendingResolverExtension.class})
+@ExtendWith({CategoryExtension.class})
 public @interface Category {
     String username();
     boolean acrhived();
