+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/extension/SpendingExtension.java
@@ -1,0 +1,53 @@
+package guru.qa.niffler.jupiter.extension;
+
+import guru.qa.niffler.api.SpendApiClient;
+import guru.qa.niffler.jupiter.annotation.Spending;
+import guru.qa.niffler.model.CategoryJson;
+import guru.qa.niffler.model.CurrencyValues;
+import guru.qa.niffler.model.SpendJson;
+import org.junit.jupiter.api.extension.*;
+import org.junit.platform.commons.support.AnnotationSupport;
+
+import java.util.Date;
+
+public class SpendingExtension implements BeforeEachCallback, ParameterResolver {
+
+    public static final ExtensionContext.Namespace NAMESPACE = ExtensionContext.Namespace.create(SpendingExtension.class);
+
+    private final SpendApiClient spendApiClient = new SpendApiClient();
+
+    @Override
+    public void beforeEach(ExtensionContext context) throws Exception {
+        AnnotationSupport.findAnnotation(context.getRequiredTestMethod(), Spending.class)
+                .ifPresent(anno -> {
+                    SpendJson spend = new SpendJson(
+                            null,
+                            new Date(),
+                            new CategoryJson(
+                                    null,
+                                    anno.category(),
+                                    anno.username(),
+                                    false
+                            ),
+                            CurrencyValues.RUB,
+                            anno.amount(),
+                            anno.description(),
+                            anno.username()
+                    );
+                    context.getStore(NAMESPACE).put(
+                            context.getUniqueId(),
+                            spendApiClient.createSpend(spend)
+                    );
+                });
+    }
+
+    @Override
+    public boolean supportsParameter(ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
+        return parameterContext.getParameter().getType().isAssignableFrom(SpendJson.class);
+    }
+
+    @Override
+    public SpendJson resolveParameter(ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
+        return extensionContext.getStore(NAMESPACE).get(extensionContext.getUniqueId(), SpendJson.class);
+    }
+}
